[Fix] change from int to uint64 allowed user id

+++ src/includes/zelibot.hpp
@@ -1,5 +1,6 @@
 #ifndef ZELIBOT_HPP
 #define ZELIBOT_HPP
+#include <cstdint>
 #include <tgbot/tgbot.h>
 #include <vector>
 
@@ -9,12 +10,12 @@ private:
   TgBot::TgLongPoll long_poll;
   const std::vector<std::string> bot_commands = {"start", "test"};
   bool test_text_state = false;
-  const int allowed_user;
+  const uint64_t allowed_user;
 
-  bool is_allowed_user(const int chat_id) const;
+  bool is_allowed_user(const uint64_t chat_id) const;
 
 public:
-  ZeliBot(const std::string &token, const int allowed_user);
+  ZeliBot(const std::string &token, const uint64_t allowed_user);
 
   void initCommands();
